Add files via upload
Kitchen Story.
Course-end Project 2

DESCRIPTION

Project objective:

Kitchen Story is an e-commerce portal that lets people shop  basic food items on their website. The website needs to have the following features:

     

A search form in the home page to allow entry of the food items to be purchased by the customer.

Based on item details entered, it will show available food items with  price.

Once a person selects an item to purchase, they will be redirected to the list of available items. In the next page, they are shown the complete breakout of the order and details of the payment to be made in the payment gateway. When payment is done, they are shown a confirmation page with details of the order.

  

For the above features to work, there will be an admin backend with the following features:

Admin login page where admin can change password after login if he wants to

A master list of food items available for purchase

A functionality to add or remove food items